Refactor to pull out handling of inputs and add function to get true output
